moved algorithms and references output of pisces category, renamed and moved data query to top of list as 'species explorer'
