+++ urls.py
@@ -10,7 +10,7 @@ import views
 import description
 import watershed
 import stream
-import query
+import species_explorer
 import algorithms
 import references
 import links_left
@@ -22,9 +22,9 @@ urlpatterns = [
     url(r'^$', description.description_page, {'model': 'pisces'}),
     url(r'^watershed$', watershed.watershed_page, {'model': 'pisces'}),
     url(r'^stream$', stream.stream_page, {'model': 'pisces'}),
-    url(r'^query$', query.query_page, {'model': 'pisces'}),
-    url(r'^algorithms$', algorithms.algorithm_page, {'model': 'pisces'}),
-    url(r'^references$', references.references_page, {'model': 'pisces'}),
+    url(r'^species$', species_explorer.query_page, {'model': 'pisces'}),
+    url(r'^algorithms/$', algorithms.algorithm_page, {'model': 'pisces'}),
+    url(r'^references/$', references.references_page, {'model': 'pisces'}),
     # url(r'^api$', rest.rest_page, {'model': 'pisces'}),
     # url(r'^swag$', views.getSwaggerJsonContent)
 ]
